reorg, move to simple cov
